central config for providers resolving
